% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_break.R
\name{take_break}
\alias{take_break}
\title{Should you take a break?}
\usage{
take_break(weather, time, workload)
}
\arguments{
\item{weather}{A character string indicating the weather.}

\item{time}{A numeric value indicating the time of day.}

\item{workload}{A numeric value indicating the amount of work you have.}
}
\value{
A logical value indicating whether you should take a break.
}
\description{
This function takes in three inputs and returns a logical value
indicating whether you should take a break.
}
\examples{
take_break(weather = "sunny", time = 10, workload = 3)


}
